Simplify lambda restatectl command and add it to cli
