+++ crates/cli-util/Cargo.toml
@@ -25,6 +25,8 @@ comfy-table = { workspace = true }
 crossterm = { workspace = true }
 dialoguer = { workspace = true }
 dotenvy = { version = "0.15" }
+lambda_runtime = "0.13.0"
+serde = { workspace = true }
 tokio = { workspace = true, features = ["time"] }
 tracing = { workspace = true }
 tracing-log = { version = "0.2.0" }
