+++ tools/restatectl/Cargo.toml
@@ -61,7 +61,6 @@ futures = { workspace = true }
 itertools = { workspace = true }
 json-patch = "2.0.0"
 humantime = { workspace = true }
-lambda_runtime = "0.13.0"
 prost-types = { workspace = true }
 rand = { workspace = true }
 rustls = { workspace = true, features = ["aws-lc-rs"]}
