+++ tools/restatectl/src/app.rs
@@ -78,6 +78,8 @@ pub enum Command {
     MetadataServer(MetadataServer),
     /// Query cluster status
     Sql(SqlOpts),
+    /// Run a AWS Lambda server
+    Lambda(restate_cli_util::lambda::LambdaOpts),
 }
 
 fn init(common_opts: &CommonOpts) {
